+++ src/web/api_handlers_recordings_backend_agnostic.c
@@ -96,9 +96,9 @@ void handle_get_recording(const http_request_t *req, http_response_t *res) {
     char size_str[32] = {0};
     if (recording.size_bytes < 1024) {
         snprintf(size_str, sizeof(size_str), "%ld B", recording.size_bytes);
-    } else if (recording.size_bytes < 1024 * 1024) {
+    } else if (recording.size_bytes < (uint64_t)1024 * 1024) {
         snprintf(size_str, sizeof(size_str), "%.1f KB", recording.size_bytes / 1024.0);
-    } else if (recording.size_bytes < 1024 * 1024 * 1024) {
+    } else if (recording.size_bytes < (uint64_t)1024 * 1024 * 1024) {
         snprintf(size_str, sizeof(size_str), "%.1f MB", recording.size_bytes / (1024.0 * 1024.0));
     } else {
         snprintf(size_str, sizeof(size_str), "%.1f GB", recording.size_bytes / (1024.0 * 1024.0 * 1024.0));
