+++ src/web/api_handlers_recordings_list.c
@@ -262,9 +262,9 @@ void handle_get_recordings(const http_request_t *req, http_response_t *res) {
         char size_str[32] = {0};
         if (recordings[i].size_bytes < 1024) {
             snprintf(size_str, sizeof(size_str), "%ld B", recordings[i].size_bytes);
-        } else if (recordings[i].size_bytes < 1024 * 1024) {
+        } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024) {
             snprintf(size_str, sizeof(size_str), "%.1f KB", recordings[i].size_bytes / 1024.0);
-        } else if (recordings[i].size_bytes < 1024 * 1024 * 1024) {
+        } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024 * 1024) {
             snprintf(size_str, sizeof(size_str), "%.1f MB", recordings[i].size_bytes / (1024.0 * 1024.0));
         } else {
             snprintf(size_str, sizeof(size_str), "%.1f GB", recordings[i].size_bytes / (1024.0 * 1024.0 * 1024.0));
